requirements.txt
langchain
langchain-openai
supabase
python-dotenv
